Optimized the menuBody.svg and improvement of code
