+++ public/js/interaction.js
@@ -1,4 +1,4 @@
-import {menuPopUp, userCoordinates, cameraCoordinates, doStateMute, doStateDeafen} from './popUpMenu.js';
+import {usePopUpMenu} from './popUpMenu.js';
 import {startSpinner} from './ClientConnection.js';
 import {peers} from './voice.js';
 
@@ -37,7 +37,7 @@ export function makeInteractable(id){
     userMove();
     userRotate();
     clickSpinner();
-    usePopUpMenu();
+    usePopUpMenu(id);
 
     return containerElement;
 
@@ -122,22 +122,5 @@ export function makeInteractable(id){
                 startSpinner();
             }
         });
-    }
-
-    // Enables the user to use the popupmenu
-    function usePopUpMenu() {
-        enableUserState();
-        userElement.onclick = (e) => menuPopUp(e, id);
-    }
-
-    function enableUserState() {
-        const muteBtn = document.getElementById("microphone");
-        const spksBtn = document.getElementById("speakers");
-        muteBtn.onclick = () => {
-            doStateMute();
-        }
-        spksBtn.onclick = () => {
-            doStateDeafen();
-        }
     }
 }
